Introduce Direction flag and optional CSV column.
Instead of having a global --bigger-is-better/--smaller-is-better setting, we
can now have a per-row setting with a new optional 'Direction' column in the CSV
format. It allows plotting different values in the same graph, such as code
size (bigger is better) and time spent (smaller is better).

To clean things up, we also introduce a --direction flag that can be set to
either 'left' or 'right' and indicates which way the biggest values should be
plotted. As a result, we have the following equivalent flags:

--bigger-is-better --> --direction=right
--smaller-is-better --> --direction=left